requirements.txt
openai==1.66.3
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
typer==0.9.0
pydantic==2.6.1
requests==2.31.0
flask==3.0.2 